+++ R/Prepare_peak.R
@@ -16,12 +16,29 @@
 #' Default: TRUE. Parameter of \code{\link{dba.count}}.
 #' @param used.cols Used columns used to create sample metadata. If specified, sampleID should be placed first. Default: c("SampleID", "Condition").
 #' @param out.folder Output folder to save created count matrix and sample metadata. Default: NULL (current working directory).
+#' @param species Species used, chosen from "Human","Mouse","Rat","Fly","Arabidopsis","Yeast","Zebrafish","Worm","Bovine","Pig","Chicken","Rhesus",
+#' "Canine","Xenopus","Anopheles","Chimp","E coli strain Sakai","Myxococcus xanthus DK 1622". Default: "Human".
+#' @param seq.style The style of sequence, chosen from UCSC, NCBI, Ensembl, None. This should be compatible with the genome and gtf file you used
+#' to generate count matrix and peak files. Default: "UCSC".
+#' @param gtf.file GTF file used to create TxDb object. Useful when specie you used is not available in \code{species}. Default: NULL.
+#' @param up.dist The upstream distance from the TSS. Default: 3000bp.
+#' @param down.dist The downstream distance from the TSS. Default: 3000bp.
+#' @param ... Parameters for \code{\link{annotatePeak}}.
 #'
-#' @return NULL
+#' @return A dataframe contains count matrix, peak annotation and sample metadata (if provided \code{used.cols}).
 #' @importFrom utils read.table write.table
 #' @importFrom magrittr %>%
-#' @importFrom dplyr mutate select
+#' @importFrom dplyr mutate select case_when
+#' @importFrom GenomicFeatures makeTxDbFromGFF
+#' @importFrom BiocManager install
+#' @importFrom GenomicRanges makeGRangesFromDataFrame
+#' @importFrom GenomeInfoDb seqlevels seqlevelsStyle keepSeqlevels seqnames
+#' @importFrom ggplotify as.ggplot
+#' @importFrom cowplot plot_grid
+#' @import ChIPseeker
+#' @import ggupset
 #' @import DiffBind
+#' @import parallel
 #'
 #' @export
 #'
@@ -30,9 +47,15 @@
 #' # library(DESeq2)
 #' # metadata file contains peak and bam information
 #' # meta.file = 'path/to/metadata'
-#' # PeakMatrix(meta.file = meta.file, blacklist = )
+#' # PeakMatrix(meta.file = meta.file, species = "Human",  seq.style = "UCSC",
+#' #            up.dist = 20000, down.dist = 20000)
 PeakMatrix <- function(meta.file, min.overlap = 2, submits = 200, use.summarizeOverlaps = TRUE,
-                       blacklist = TRUE, sub.control = TRUE, used.cols = c("SampleID", "Condition"), out.folder = NULL) {
+                       blacklist = TRUE, sub.control = TRUE, used.cols = c("SampleID", "Condition"),
+                       out.folder = NULL, species = c(
+                         "Human", "Mouse", "Rat", "Fly", "Arabidopsis", "Yeast", "Zebrafish", "Worm", "Bovine", "Pig", "Chicken", "Rhesus",
+                         "Canine", "Xenopus", "Anopheles", "Chimp", "E coli strain Sakai", "Myxococcus xanthus DK 1622"
+                       ),
+                       seq.style = c("UCSC", "NCBI", "Ensembl", "None"), gtf.file = NULL, up.dist = 3000, down.dist = 3000, ...) {
   # load sample meta info
   sample.info <- read.table(meta.file, sep = "\t", header = TRUE, check.names = FALSE)
   # check columns
@@ -43,6 +66,7 @@ PeakMatrix <- function(meta.file, min.overlap = 2, submits = 200, use.summarizeO
   # create dba object
   dbaObj <- DiffBind::dba(sampleSheet = sample.info, minOverlap = min.overlap)
   # count
+  suppressWarnings(suppressMessages(library(parallel)))
   dbaObj <- DiffBind::dba.count(dbaObj, bUseSummarizeOverlaps = use.summarizeOverlaps)
   # remove blacklist region
   if (!is.null(blacklist)) {
@@ -65,12 +89,52 @@ PeakMatrix <- function(meta.file, min.overlap = 2, submits = 200, use.summarizeO
     dbaObj <- DiffBind::dba.count(dbaObj, peaks = NULL, score = DBA_SCORE_READS)
   }
   count.df <- DiffBind::dba.peakset(dbaObj, bRetrieve = TRUE, DataType = DBA_DATA_FRAME)
-  # merge columns
+  # prepare peak dataframe for peak annotation
+  peak.df <- count.df[c("CHR", "START", "END")]
+  colnames(peak.df) <- c("chr", "start", "stop")
+  peak.df$name <- paste(peak.df$chr, paste(peak.df$start, peak.df$stop, sep = "-"), sep = ":")
+  # peak annotation
+  peak.anno <- AnnoPeak(
+    peak.df = peak.df, species = species, seq.style = seq.style,
+    gtf.file = gtf.file, up.dist = up.dist, down.dist = down.dist, ...
+  )
+  peak.anno.df <- peak.anno$df
+  # change anno type
+  peak.anno.df <- peak.anno.df %>%
+    dplyr::mutate(AnnoType = dplyr::case_when(
+      anno == "Promoter" ~ "P",
+      anno == "5' UTR" ~ "5U",
+      anno == "3' UTR" ~ "3U",
+      anno == "Exon" ~ "E",
+      anno == "Intron" ~ "I",
+      anno == "Downstream" ~ "D",
+      anno == "Distal Intergenic" ~ "DI"
+    )) %>%
+    dplyr::mutate(FullName = paste(name, SYMBOL, AnnoType, sep = "|")) %>%
+    dplyr::select(c("name", "FullName"))
+  # merge count columns
   count.df <- count.df %>%
-    dplyr::mutate(feature = paste(CHR, paste(START, END, sep = "-"), sep = ":")) %>%
+    dplyr::mutate(name = paste(CHR, paste(START, END, sep = "-"), sep = ":")) %>%
     dplyr::select(-c(CHR, START, END))
-  rownames(count.df) <- count.df$feature
-  count.df$feature <- NULL
+  # merge paek annotation and count
+  count.df <- merge(count.df, peak.anno.df, by = "name", all.x = TRUE)
+  rownames(count.df) <- count.df$FullName
+  count.df$name <- NULL
+  count.df$FullName <- NULL
+  # save data
+  if (is.null(out.folder)) {
+    out.folder <- getwd()
+  }
+  # write count matrix
+  write.table(
+    x = count.df, file = file.path(out.folder, "consensus_peak_matrix.txt"),
+    sep = "\t", quote = FALSE, col.names = TRUE
+  )
+  # write peak annotation results
+  write.table(
+    x = peak.anno$df, file = file.path(out.folder, "consensus_peak_anno.txt"),
+    sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE
+  )
   # create metadata
   if (!is.null(used.cols)) {
     valid.used.cols <- intersect(colnames(sample.info), used.cols)
@@ -85,14 +149,8 @@ PeakMatrix <- function(meta.file, min.overlap = 2, submits = 200, use.summarizeO
       x = sample.meta, file = file.path(out.folder, "peak_metadata.txt"),
       sep = "\t", quote = FALSE, row.names = TRUE, col.names = TRUE
     )
+    return(list(count = count.df, meta = sample.meta, peak_anno = peak.anno$df))
+  } else {
+    return(list(count = count.df, peak_anno = peak.anno$df))
   }
-  # save data
-  if (is.null(out.folder)) {
-    out.folder <- getwd()
-  }
-  # write count matrix
-  write.table(
-    x = count.df, file = file.path(out.folder, "consensus_peak_matrix.txt"),
-    sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE
-  )
 }